Логика по поиску соединений помещений перенесена в отдельный класс
