+++ src/RevitDeclarations/Models/UtpCalculator.cs
@@ -16,23 +16,15 @@ namespace RevitDeclarations.Models {
         private const int _minPantryDepth = 1000;
         private const double _minPantryArea = 1.8;
 
-        private readonly StringComparer _strComparer = StringComparer.OrdinalIgnoreCase;
-        private readonly StringComparison _strComparison = StringComparison.OrdinalIgnoreCase;
-
         private readonly ApartmentsProject _project;
         private readonly ApartmentsSettings _settings;
         private readonly PrioritiesConfig _priorities;
+        private readonly RoomConnectionAnalyzer _roomConnectionAnalyzer;
 
         private readonly bool _hasNullAreas;
         private readonly bool _hasBannedNames;
         private readonly IEnumerable<ElementId> _roomsWithBathFamily;
 
-        // Гардеробные с дверью в жилую комнату. Для УТП Гардеробная
-        private IEnumerable<ElementId> _pantriesWithBedroom;
-        // Мастер-спальни 
-        private IEnumerable<ElementId> _masterBedrooms;
-        // Санузлы, относящиеся к мастер спальням
-        private IEnumerable<ElementId> _masterBathrooms;
 
         public UtpCalculator(ApartmentsProject project, DeclarationSettings settings) {
             _project = project;
@@ -43,40 +35,12 @@ namespace RevitDeclarations.Models {
             _roomsWithBathFamily = GetRoomsWithBath();
 
             _priorities = _settings.PrioritiesConfig;
+
+            _roomConnectionAnalyzer = new RoomConnectionAnalyzer(project, _settings);
         }
 
         public void CalculateRoomsForUtp() {
-            var bedroomBathroom = GetRoomsByDoors(_priorities.LivingRoom, _priorities.Bathroom);
-            // Жилые комната, соединенные с санузлами
-            List<ElementId> bedroomsWithBathroom = bedroomBathroom[_priorities.LivingRoom.Name];
-
-            var pantryBedroom = GetRoomsByDoors(_priorities.Pantry, _priorities.LivingRoom);
-            var pantryBathroom = GetRoomsByDoors(_priorities.Pantry, _priorities.Bathroom);
-            // Гардеробные, соединенные с жилыми комнатами (не учитываются для УТП Гардеробная)
-            _pantriesWithBedroom = pantryBedroom[_priorities.Pantry.Name];
-            // Гардеробные, соединенные с санузлами
-            List<ElementId> pantriesWithBathroom = pantryBathroom[_priorities.Pantry.Name];
-            // Мастер-гардеробные (имеющие дверь в санузел) для определения мастер-спален
-            List<ElementId> masterPantries = _pantriesWithBedroom.Intersect(pantriesWithBathroom).ToList();
-
-            var bedroomPantry = GetRoomsByDoors(_priorities.LivingRoom, masterPantries);
-            // Жилые комнаты, соединенные с мастер-гардеробными
-            var bedroomsWithPantryAndBathroom = bedroomPantry[_priorities.LivingRoom.Name];
-
-            // Итоговый список мастер-спален
-            _masterBedrooms = bedroomsWithBathroom
-                .Concat(bedroomsWithPantryAndBathroom)
-                .ToList();
-
-            List<ElementId> bathroomsWithBedroom = bedroomBathroom[_priorities.Bathroom.Name];
-            var bathroomPantry = GetRoomsByDoors(_priorities.Bathroom, masterPantries);
-            // Санузлы, соединенные с мастер-гардеробными
-            var bathroomsWithMasterPantry = bathroomPantry[_priorities.Bathroom.Name];
-
-            // Санузлы, относящиеся к мастер-спальням для определение УТП Две ванны
-            _masterBathrooms = bathroomsWithBedroom
-                .Concat(bathroomsWithMasterPantry)
-                .ToList();
+            _roomConnectionAnalyzer.FindConnections();
         }
 
         public IReadOnlyCollection<ErrorsListViewModel> CheckProjectForUtp() {
@@ -145,7 +109,7 @@ namespace RevitDeclarations.Models {
 
             bathAmount = apartment.Rooms
                 .Select(x => x.RevitRoom.Id)
-                .Where(x => !_masterBathrooms.Contains(x))
+                .Where(x => !_roomConnectionAnalyzer.CheckIsMasterBathroom(x))
                 .Where(x => _roomsWithBathFamily.Contains(x))
                 .Count();
 
@@ -191,7 +155,7 @@ namespace RevitDeclarations.Models {
             }
 
             foreach(var room in apartment.Rooms) {
-                if(_masterBedrooms.Contains(room.RevitRoom.Id)) {
+                if(_roomConnectionAnalyzer.CheckIsMasterBedroom(room.RevitRoom.Id)) {
                     return "Да";
                 }
             }
@@ -246,7 +210,7 @@ namespace RevitDeclarations.Models {
             List<Room> pantriesWithoutBedrooms = apartment
                 .GetRoomsByPrior(_priorities.Pantry)
                 .Select(x => x.RevitRoom)
-                .Where(x => !_pantriesWithBedroom.Contains(x.Id))
+                .Where(x => !_roomConnectionAnalyzer.CheckIsPantryWithBedroom(x.Id))
                 .Where(x => ContourChecker.CheckArea(x, _settings.AccuracyForArea, _minPantryArea))
                 .ToList();
 
@@ -287,77 +251,6 @@ namespace RevitDeclarations.Models {
                 .Where(x => _settings.BannedRoomNames.Contains(x, StringComparer.OrdinalIgnoreCase))
                 .Distinct()
                 .ToList();
-        }
-
-        private Dictionary<string, List<ElementId>> GetRoomsByDoors(RoomPriority priority1, RoomPriority priority2) {
-            string name1 = priority1.Name;
-            string name2 = priority2.Name;
-            BuiltInParameter bltParam = BuiltInParameter.ROOM_NAME;
-            IReadOnlyCollection<FamilyInstance> doors = _project.GetDoors();
-
-            Dictionary<string, List<ElementId>> roomByNames = new Dictionary<string, List<ElementId>>(_strComparer) {
-                [name1] = new List<ElementId>(),
-                [name2] = new List<ElementId>()
-            };
-
-            foreach(var door in doors) {
-                Room room1 = door.get_FromRoom(_project.Phase);
-                Room room2 = door.get_ToRoom(_project.Phase);
-                // У дверей может не быть помещения с одной стороны
-                if(room1 != null && room2 != null) { 
-                    string roomName1 = room1.get_Parameter(bltParam).AsString().ToLower();
-                    string roomName2 = room2.get_Parameter(bltParam).AsString().ToLower();
-
-                    if(string.Equals(roomName1, name1, _strComparison) && 
-                       string.Equals(roomName2, name2, _strComparison)) {
-                        roomByNames[name1].Add(room1.Id);
-                        roomByNames[name2].Add(room2.Id);
-                    }
-
-                    if(string.Equals(roomName1, name2, _strComparison) && 
-                       string.Equals(roomName2, name1, _strComparison)) {
-                        roomByNames[name1].Add(room2.Id);
-                        roomByNames[name2].Add(room1.Id);
-                    }                
-                }
-            }
-
-            return roomByNames;
-        }
-
-        private Dictionary<string, List<ElementId>> GetRoomsByDoors(RoomPriority priority1, List<ElementId> rooms) {
-            string name1 = priority1.Name;
-            BuiltInParameter bltParam = BuiltInParameter.ROOM_NAME;
-            IReadOnlyCollection<FamilyInstance> doors = _project.GetDoors();
-
-            Dictionary<string, List<ElementId>> roomByNames = new Dictionary<string, List<ElementId>>(_strComparer) {
-                [name1] = new List<ElementId>(),
-            };
-
-            foreach(var door in doors) {
-                Room room1 = door.get_FromRoom(_project.Phase);
-                Room room2 = door.get_ToRoom(_project.Phase);
-                // У дверей может не быть помещения с одной стороны
-                if(room1 != null && room2 != null) {
-                    string roomName1 = room1.get_Parameter(bltParam).AsString().ToLower();
-                    string roomName2 = room2.get_Parameter(bltParam).AsString().ToLower();
-
-                    if(string.Equals(roomName1, name1, _strComparison) && rooms.Contains(room2.Id)) {
-                        roomByNames[name1].Add(room1.Id);
-                    }
-
-                    if(string.Equals(roomName2, name1, _strComparison) && rooms.Contains(room1.Id)) {
-                        roomByNames[name1].Add(room2.Id);
-                    }
-                }
-            }
-
-            return roomByNames;
-        }
-
-        private Dictionary<string, List<ElementId>> GetRoomsBySeparators(RoomPriority priority1, 
-                                                                         RoomPriority priority2) {
-
         }
 
         private IEnumerable<ElementId> GetRoomsWithBath() {
